댓글삭제여부

+++ TM_Front/src/comments/pages/CommentList.tsx
@@ -83,6 +83,9 @@ export default function CommentList({
   /** 댓글 삭제 */
   const handleDelete = () => {
     if (selectedId !== null) {
+      const confirmed = window.confirm("해당 댓글을 삭제하시겠습니까?");
+      if (!confirmed) return;
+
       onDelete(selectedId);
       handleClose();
     }
